crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = ""
authors = [
    {name = "ru1713",email = "konishi2039@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp (>=3.12.9,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
